# Conflicts:
#	wdmovie/build.gradle
#	wdmovie/src/main/AndroidManifest.xml

+++ wdmovie/src/main/java/com/bw/movie/camera/bean/FindMovieListByCinemaIdBean.java
@@ -57,18 +57,18 @@ public class FindMovieListByCinemaIdBean {
          * summary : 故事发生在《夏洛特烦恼》中的“特烦恼”之城“西虹市”。混迹于丙级业余足球队的守门员王多鱼（沈腾 饰），因比赛失利被开除离队。正处于人生最低谷的他接受了神秘台湾财团“一个月花光十亿资金”的挑战。本以为快乐生活就此开始，王多鱼却第一次感到“花钱特烦恼”！想要人生反转走上巅峰，真的没有那么简单.
          */
 
-        private int fare;
+        private double fare;
         private int id;
         private String imageUrl;
         private String name;
         private long releaseTime;
         private String summary;
 
-        public int getFare() {
+        public double getFare() {
             return fare;
         }
 
-        public void setFare(int fare) {
+        public void setFare(double fare) {
             this.fare = fare;
         }
